refactor: TransactionForm komponens használata
