+++ src/ExpenseApp/ExpenseApp.jsx
@@ -1,25 +1,11 @@
 import React, { useState } from "react";
+import TransactionForm from "./components/TransactionForm";
 
 function ExpenseApp() {
   const [transactions, setTransactions] = useState([]);
-  const [description, setDescription] = useState("");
-  const [amount, setAmount] = useState("");
-  const [type, setType] = useState("income");
 
-  const addTransaction = () => {
-    if (!description || !amount) return;
-
-    const newTransaction = {
-      id: Date.now(),
-      description,
-      amount: parseFloat(amount),
-      type,
-    };
-
-    setTransactions([...transactions, newTransaction]);
-    setDescription("");
-    setAmount("");
-    setType("income");
+  const addTransaction = (data) => {
+    setTransactions([...transactions, data]);
   };
 
   const income = transactions.filter((t) => t.type === "income");
@@ -29,27 +15,9 @@ function ExpenseApp() {
     <div>
       <h1>Havi Költségkezelő</h1>
 
+      <TransactionForm addTransaction={(d) => addTransaction(d)} />
+
       <div>
-        <select value={type} onChange={(e) => setType(e.target.value)}>
-          <option value="income">+</option>
-          <option value="expense">-</option>
-        </select>
-
-        <input
-          type="text"
-          placeholder="Leírás"
-          value={description}
-          onChange={(e) => setDescription(e.target.value)}
-        />
-
-        <input
-          type="number"
-          placeholder="Összeg"
-          value={amount}
-          onChange={(e) => setAmount(e.target.value)}
-        />
-        <button onClick={addTransaction}>Hozzáadás</button>
-
         <div>
           <h3>Bevételek</h3>
           <ul>
